规范了代码格式

+++ WebContent/WEB-INF/jsp/admin/buser/selectBuser.jsp
@@ -14,7 +14,9 @@
 <title>Insert title here</title>
 <link href="css/admin.css" type="text/css" rel="stylesheet">
 <script src="js/colorControl.js" type="text/javascript"></script>
+<script src="js/buttonAction.js" type="text/javascript"></script>
 <script src="js/formCheck.js" type="text/javascript"></script>
+<script src="js/postConfirm.js" type="text/javascript"></script>
 </head>
 <body>
 	<form:form id="searchForm" action="adminBuser/searchBuser"
@@ -46,7 +48,7 @@
 					<td>${buser.uname }</td>
 					<td>${buser.uidnum }</td>
 					<td>${buser.upwd }</td>
-					<td><a onclick="return checkDel();"
+					<td><a onclick="return confirmDel();"
 						href="adminBuser/deleteBuser?user_id=${buser.user_id }">删除</a></td>
 				</tr>
 			</c:forEach>
